requirements.txt
albumentations==1.4.4
annotated-types==0.6.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.9.0
fonttools==4.51.0
fsspec==2023.4.0
imageio==2.34.0
Jinja2==3.1.2
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.3
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
typing_extensions==4.11.0
